requirements.txt
Django==2.0.7
mysqlclient==1.3.13
pytz==2018.5
djangorestframework==3.8.2
